Move(userAuth): exception 경로 이동
