+++ src/main/java/com/sparta/newsfeed/user/exception/ExistingUserException.java
@@ -1,4 +1,4 @@
-package com.sparta.newsfeed.user;
+package com.sparta.newsfeed.user.exception;
 
 public class ExistingUserException extends IllegalArgumentException{
