+++ src/main/java/com/sparta/newsfeed/user/exception/WrongPasswordException.java
@@ -1,4 +1,4 @@
-package com.sparta.newsfeed.user;
+package com.sparta.newsfeed.user.exception;
 
 public class WrongPasswordException extends IllegalArgumentException{
     private static final String MESSAGE = "비밀번호가 일치하지 않습니다.";
