--- src/main/resources/application.properties
@@ -1,8 +1,0 @@
-# DB ??
-spring.profiles.include=secret
-
-## hibernate ??
-spring.jpa.hibernate.ddl-auto=update
-spring.jpa.properties.hibernate.show_sql=true
-spring.jpa.properties.hibernate.format_sql=true
-spring.jpa.properties.hibernate.use_sql_comments=true
